Refactor WFS API to be more easily extendable
- Move common WFS classes to `traffic-control` app.
- Change WFS Endpoint root url to `/wfs/`
