+++ city_furniture/views/wfs/__init__.py
@@ -1,0 +1,2 @@
+# flake8: noqa
+from .furniture_signpost import FurnitureSignpostPlanFeatureType, FurnitureSignpostRealFeatureType
